commit30012020

+++ Controllers/StudentController.cs
@@ -63,7 +63,12 @@ namespace CodeFirstEFCore.Controllers
 
         [HttpDelete("{id}")]
         public ActionResult Delete(int id)
-        {
+        {    
+            if (!ModelState.IsValid)
+            {
+                return BadRequest();
+            }
+
             _studentService.delete(id);
             return Ok();
         }
